+++ packages/react/src/hooks/banksocial/index.ts
@@ -1,2 +1,0 @@
-export { useBankSocialActivity } from './useBankSocialActivity'
-export { useCreateDAO } from './useCreateDAO'