+++ packages/react/src/utils/constants/addresses.ts
@@ -1,4 +1,17 @@
+// polygon mainnet
 export const socialBankAddress = '0x42da2051704c3fa61637a4b80b73ee9be11bdc46'
-export const daoAddress = '0x4b11b1234dE09d72e59AcC2f089A746337566Deb'
-export const daoVaultAddress = '0xf997caC57F39E3cE8C13bcB01ddE9fE1209CB4B4'
-export const memberCardAddress = '0xf9d2CD13502d4FC8B105057e23c83D23AE43B0Ea'
+export const daoAddress = '0xF8242a819d1481955FEEe1Bf76F0F9E785911D09'
+export const daoVaultAddress = '0xa3989e09582EDA3048CD1372F7fE5Bb97eBc5Bc8'
+export const memberCardAddress = '0xCA5DdbbE67d21b52aA8b50Cd2E46dE246D0B1771'
+export const usdcAddress = '0x2791Bca1f2de4661ED88A30C99A7a9449Aa84174'
+export const aaveATokenAddress = '0x625E7708f30cA75bfd92586e17077590C60eb4cD'
+export const poolAddress = '0xa97684ead0e402dC232d5A977953DF7ECBaB3CDb'
+export const uniswapAddress = '0xC7cF646eDB2425613624adD90C4fAA1cf86136E1'
+
+// mumbai testnet
+export const mumbaiSocialBankAdr = '0xE549D25C8037aad1b3aC32d444eB1a6f82ca9DAB'
+export const mumbaiDaoAdr = '0x4b11b1234dE09d72e59AcC2f089A746337566Deb'
+export const mumbaiDaoVaultAdr = '0xf997caC57F39E3cE8C13bcB01ddE9fE1209CB4B4'
+export const mumbaiMemberCardAdr = '0xf9d2CD13502d4FC8B105057e23c83D23AE43B0Ea'
+
+// https://polygonscan.com/address/0x42da2051704c3fa61637a4b80b73ee9be11bdc46#code