+++ packages/react/src/utils/constants/index.ts
@@ -1,7 +1,0 @@
-export { socialBankABI, daoABI, daoVaultABI, memberCardABI } from './abis'
-export {
-  socialBankAddress,
-  daoAddress,
-  daoVaultAddress,
-  memberCardAddress,
-} from './addresses'