+++ packages/react/src/utils/index.ts
@@ -1,13 +1,2 @@
 export { deserialize } from './deserialize'
 export { serialize } from './serialize'
-
-export {
-  socialBankAddress,
-  socialBankABI,
-  daoABI,
-  daoAddress,
-  daoVaultABI,
-  daoVaultAddress,
-  memberCardABI,
-  memberCardAddress,
-} from './constants'