+++ examples/_dev/CHANGELOG.md
@@ -1,5 +1,14 @@
 # example-dev
 
+## 0.0.6
+
+### Patch Changes
+
+- add contract calls
+
+- Updated dependencies []:
+  - wagmi-banksocial@0.7.14
+
 ## 0.0.5
 
 ### Patch Changes